# src/main/resources/openapi/madkursus-api.yaml
openapi: 3.0.3
info:
  title: Madkursus API
  version: v1
  description: API for managing products and inventory.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://madkursus.jamesbabz.duckdns.org
    description: Production
...
paths:
...
  /v1/inventory:
    get:
      tags: [inventory]
      operationId: getInventory
      responses:
        '200':
          description: Inventory items
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/InventoryItem'}
        '500': {$ref: '#/components/responses/serverException'}
    post:
      tags: [inventory]
      operationId: createInventoryItem
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AddInventoryItem'}
      responses:
        '201':
          description: Inventory item created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InventoryItem'}
        '400': {$ref: '#/components/responses/badRequest'}
        '404': {$ref: '#/components/responses/notFound'}
        '500': {$ref: '#/components/responses/serverException'}
  /v1/inventory/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags: [inventory]
      operationId: getInventoryItem
      responses:
        '200':
          description: Inventory item
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InventoryItem'}
        '404': {$ref: '#/components/responses/notFound'}
        '500': {$ref: '#/components/responses/serverException'}
    patch:
      tags: [inventory]
      operationId: updateInventoryItem
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateInventoryItem'}
      responses:
        '204': {description: 'Inventory amount updated, or row removed when quantity is zero'}
        '400': {$ref: '#/components/responses/badRequest'}
        '404': {$ref: '#/components/responses/notFound'}
        '500': {$ref: '#/components/responses/serverException'}
    delete:
      tags: [inventory]
      operationId: deleteInventoryItem
      responses:
        '204': {description: Inventory item deleted}
        '404': {$ref: '#/components/responses/notFound'}
        '500': {$ref: '#/components/responses/serverException'}
  /v1/inventory/from-template/{templateId}:
    parameters:
      - name: templateId
        in: path
        required: true
        schema: {type: string, format: uuid}
    post:
      tags: [inventory]
      operationId: createInventoryItemFromTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AddInventoryQuantity'}
      responses:
        '201':
          description: Inventory stock added from template
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InventoryItem'}
        '400': {$ref: '#/components/responses/badRequest'}
        '404': {$ref: '#/components/responses/notFound'}
...
  /v1/kitchen-equipment:
    get:
      tags: [kitchen-equipment]
      operationId: getKitchenEquipmentList
      responses:
        '200': {description: Current user's kitchen equipment, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/KitchenEquipment'}}}}}
        '401': {$ref: '#/components/responses/unauthorized'}
    post:
      tags: [kitchen-equipment]
      operationId: createKitchenEquipment
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/KitchenEquipmentInput'}}}}
      responses:
        '201': {description: Equipment created, content: {application/json: {schema: {$ref: '#/components/schemas/KitchenEquipment'}}}}
        '400': {$ref: '#/components/responses/badRequest'}
        '409': {$ref: '#/components/responses/conflict'}
  /v1/kitchen-equipment/{id}:
    parameters: [{$ref: '#/components/parameters/id'}]
    get:
      tags: [kitchen-equipment]
      operationId: getKitchenEquipment
      responses:
        '200': {description: Equipment, content: {application/json: {schema: {$ref: '#/components/schemas/KitchenEquipment'}}}}
        '404': {$ref: '#/components/responses/notFound'}
    patch:
      tags: [kitchen-equipment]
      operationId: updateKitchenEquipment
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/KitchenEquipmentInput'}}}}
      responses:
        '200': {description: Equipment updated, content: {application/json: {schema: {$ref: '#/components/schemas/KitchenEquipment'}}}}
        '400': {$ref: '#/components/responses/badRequest'}
        '404': {$ref: '#/components/responses/notFound'}
        '409': {$ref: '#/components/responses/conflict'}
    delete:
      tags: [kitchen-equipment]
      operationId: deleteKitchenEquipment
      responses:
        '204': {description: Equipment deleted}
        '404': {$ref: '#/components/responses/notFound'}
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
  responses:
    badRequest:
      description: Invalid request
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorMessage'}
    notFound:
      description: Resource not found
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorMessage'}
    conflict:
      description: Request conflicts with current state
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorMessage'}
    unauthorized:
      description: Authentication required or credentials invalid
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorMessage'}
...
    serverException:
      description: Unexpected server error
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorMessage'}
  schemas:
    EquipmentType:
      type: string
      enum: [STOVE, OVEN, POT, PAN, AIR_FRYER, THERMOMETER, MICROWAVE]
    HeatSource:
      type: string
      enum: [INDUCTION, CERAMIC, ELECTRIC, GAS, OTHER]
    OvenMode:
      type: string
      enum: [CONVENTIONAL, FAN, GRILL]
    ThermometerType:
      type: string
      enum: [INSTANT_READ, PROBE, OTHER]
    KitchenEquipmentInput:
      type: object
      required: [equipmentType, name]
      properties:
        equipmentType: {$ref: '#/components/schemas/EquipmentType'}
        name: {type: string, minLength: 1, maxLength: 255}
        active: {type: boolean, default: true}
        preferred: {type: boolean, default: false}
        heatSource: {$ref: '#/components/schemas/HeatSource'}
        minimumLevel: {type: integer, minimum: 0, nullable: true}
        maximumLevel: {type: integer, minimum: 0, nullable: true}
        heatMappings:
          type: object
          nullable: true
          additionalProperties: {type: string}
        ovenModes:
          type: array
          uniqueItems: true
          items: {$ref: '#/components/schemas/OvenMode'}
        minimumTemperatureCelsius: {type: integer, minimum: 1, nullable: true}
        maximumTemperatureCelsius: {type: integer, minimum: 1, nullable: true}
        capacityMl: {type: integer, minimum: 1, nullable: true}
        diameterMm: {type: integer, minimum: 1, nullable: true}
        nonStick: {type: boolean, nullable: true}
        thermometerType: {$ref: '#/components/schemas/ThermometerType'}
        maxPowerWatts: {type: integer, minimum: 1, nullable: true}
    KitchenEquipment:
      allOf:
        - $ref: '#/components/schemas/KitchenEquipmentInput'
        - type: object
          required: [id, active, preferred, createdAt, updatedAt]
          properties:
            id: {type: string, format: uuid, readOnly: true}
            createdAt: {type: string, format: date-time, readOnly: true}
            updatedAt: {type: string, format: date-time, readOnly: true}
...
    ProductCategory:
      type: string
      enum: [BAKING, BREAD, DAIRY, EGG, FISH, FROZEN, FRUIT, GRAIN_PASTA, HERB, LEGUME, MEAT, NUT_SEED, OIL_FAT, OTHER, PRESERVED, SAUCE_CONDIMENT, SPICE, STOCK, SWEETENER, VEGETABLE, VINEGAR_ACID, DRY_GOODS]
    Unit:
      type: string
      enum: [GRAM, MILLILITER, PIECE]
    InventoryTrackingMode:
      type: string
      enum: [QUANTITY, PRESENCE]
    CreateProduct:
      type: object
      required: [name, category, defaultUnit]
      properties:
        name: {type: string, minLength: 1, example: Løg}
        category: {$ref: '#/components/schemas/ProductCategory'}
        defaultUnit: {$ref: '#/components/schemas/Unit'}
        inventoryTrackingMode: {$ref: '#/components/schemas/InventoryTrackingMode'}
    Product:
      allOf:
        - $ref: '#/components/schemas/CreateProduct'
        - type: object
          required: [id, inventoryTrackingMode]
          properties:
            id: {type: string, format: uuid}
            sourceTemplateId: {type: string, format: uuid, nullable: true, readOnly: true}
...
    AddInventoryItem:
      type: object
      required: [productId]
      properties:
        productId: {type: string, format: uuid}
        quantity: {type: number, minimum: 0, exclusiveMinimum: true, multipleOf: 0.5, nullable: true, example: 500}
    AddInventoryQuantity:
      type: object
      properties:
        quantity: {type: number, minimum: 0, exclusiveMinimum: true, multipleOf: 0.5, nullable: true, example: 10}
    UpdateInventoryItem:
      type: object
      required: [quantity]
      properties:
        quantity: {type: number, minimum: 0, multipleOf: 0.5, example: 800}
    InventoryItem:
      type: object
      required: [id, product, unit, available, plannedUsageCount, reservations]
      properties:
        id: {type: string, format: uuid}
        product: {$ref: '#/components/schemas/Product'}
        quantity: {type: number, minimum: 0, multipleOf: 0.5, nullable: true}
        physicalQuantity: {type: number, minimum: 0, nullable: true, readOnly: true}
        reservedQuantity: {type: number, minimum: 0, nullable: true, readOnly: true}
        availableQuantity: {type: number, minimum: 0, nullable: true, readOnly: true}
        plannedShortfall: {type: number, minimum: 0, nullable: true, readOnly: true}
        plannedUsageCount: {type: integer, minimum: 0, readOnly: true}
        reservations: {type: array, readOnly: true, items: {$ref: '#/components/schemas/InventoryReservationDetail'}}
        unit: {$ref: '#/components/schemas/Unit'}
        available: {type: boolean}
    InventoryReservationDetail:
      type: object
      required: [mealPlanId, mealPlanName, plannedRecipeId, recipeId, recipeName, portions, unit]
      properties:
        mealPlanId: {type: string, format: uuid}
        mealPlanName: {type: string}
        plannedRecipeId: {type: string, format: uuid}
        recipeId: {type: string, format: uuid}
        recipeName: {type: string}
        portions: {type: integer, minimum: 1}
        reservedQuantity: {type: number, minimum: 0, nullable: true}
        unit: {$ref: '#/components/schemas/Unit'}
...
    ErrorMessage:
      type: object
      required: [message, errors, timestamp, correlationId]
      properties:
        message: {type: string}
        errors:
          type: array
          items: {type: string}
        timestamp: {type: string, format: date-time}
        correlationId: {type: string, format: uuid}